Added daemonize function and cleaned up code.
Some basic daemon functions were added.
daemonize() will run the server as a daemon.
is_daemon() will check to see if the server is a daemon.

The code was cleaned up to get rid of some echo statements where there
didn't need to be any. The variable defining the data directory is in
bootstrap.php. The $data_dir variable will be used from here on out to
deal with things like server logs, configuration data, the server pid if
running as a daemon, etc.

The server_log() function was changed to check if the server is a
daemon, and if so, will not output to the console as that would cause a crash.

for your changes. Lines starting